+++ api/controllers/api/admin/RoleDetailController.js
@@ -64,13 +64,16 @@ module.exports = {
 
   update: async (req, res) => {
     try {
-      const { id } = req.params;
+      const { id, MenuItemId } = req.params;
       const data = req.body;
+      const findMenuItemId = await MenuItem.findOne({ where: { id: data.MenuItemId } });
+      const loinUser = AuthService.getSessionUser(req).username;
       const message = 'Update success.';
       let checkRoleDetailHaveREADName = await RoleDetailService.checkRoleDetailHaveREADName({ data });
       const checkHaveSameRole = await RoleDetailService.checkHaveSameRole({ data });
-      checkRoleDetailHaveREADName = checkRoleDetailHaveREADName || data.name === 'READ' || data.name === 'READ_WRITE'; 
-      const checkSuccess = checkRoleDetailHaveREADName && !checkHaveSameRole;
+      checkRoleDetailHaveREADName = checkRoleDetailHaveREADName || data.name === 'READ' || data.name === 'READ_WRITE';
+      const checkAdminUpdateRolePage = (findMenuItemId.dataValues.title === '詳細權限' || findMenuItemId.dataValues.title === '權限') && loinUser === 'admin';
+      const checkSuccess = checkRoleDetailHaveREADName && !checkHaveSameRole && !checkAdminUpdateRolePage;
       if(checkSuccess) {
         const item = await RoleDetail.update(data ,{
           where: { id, },
@@ -79,6 +82,8 @@ module.exports = {
       } else {
         if(checkHaveSameRole) {
           res.serverError('此權限已存在!');
+        } else if(checkAdminUpdateRolePage) {
+          res.serverError("不能修改自己的權限");
         } else {
           res.serverError('需要先有 READ 權限才能更新此權限!');
         }
@@ -91,9 +96,17 @@ module.exports = {
   destroy: async (req, res) => {
     try {
       const { id } = req.params;
-      const item = await RoleDetail.destroy({ where: { id } });
-      let message = 'Delete success';
-      res.ok({message, data: {item}});
+      const MenuItemId = await  RoleDetail.findOne({ where: { id } });
+      const findMenuItemId = await MenuItem.findOne({ where: { id: MenuItemId.dataValues.MenuItemId } });
+      const loinUser = AuthService.getSessionUser(req).username;
+      const checkAdminUpdateRolePage = (findMenuItemId.dataValues.title === '詳細權限' || findMenuItemId.dataValues.title === '權限') && loinUser === 'admin';
+      if(checkAdminUpdateRolePage) {
+        res.serverError("不能刪除自己的權限");
+      } else {
+        const item = await RoleDetail.destroy({ where: { id } });
+        let message = 'Delete success';
+        res.ok({message, data: {item}});
+      }
     } catch (e) {
       res.serverError(e);
     }
